Enable love-minor-mode if we find any built-ins in a buffer
This patch changes love/automatically-enable to perform a search for
any of the built-in callback functions and standard modules in LÖVE.
If we find them then we enable the mode.  This works more effectively
(thus far) than searching against file names, as it is less likely to
produce false positives.