Implement support for opening online and offline documentation
This patch adds two variables users can customize:

1. The URL to the official LÖVE wiki, which we provide by default.

2. A path to a copy of local documentation which is available from the
LÖVE website, which is an empty string by default.

Then we define a function for opening the user’s browser with that
documentation, preferring local documentation to the online wiki.
This is because I personally work offline a lot, and therefore prefer
to open my local copy most of the time.  We bind this command to the
key sequence ‘C-c C-f’.