Do not move forward in the buffer when enabling the mode
The function love/automatically-enable calls re-search-forward to
decide whether or not to enable the minor mode.  This has the
undesired side-effect of moving the point.  So if we do find something
we want to move the point back to the beginning of the buffer so that
users are not opening a file and finding themselves unexpectedly in
the middle of it.