Change the source comments to the standard Emacs package format
This patch contains no logic changes.  All it does is modify the
format of the main source header comment so that it conforms to the
standard seen in most Emacs packages.  This makes it possible to
submit love-minor-mode to the package repositories out there, as they
look for metadata from comments in this format.

This commit also updates the copyright message for 2013.