Rewrite love/automatically-enable to use save-excursion
This is a better approach since the function uses goto-char to jump to
the beginning of the buffer to search for LÖVE functions.  There is a
chance a user may open a Lua file through a bookmarked point, for
example, so in that case we would not want the existing behavior of
moving the point to the start of the buffer.  By using save-excursion
we can avoid this potential problem.