[Fix] Try to use the default path to the `love` executable
Emacs has an `executable-find` function which accepts the name of a
command and searches `exec-path` for the full path to that command.
Basically, it searches `$PATH` like many shells.  We can use...

    (executable-find "love")

...to find the path to that program, e.g. `/usr/local/bin/love`.  By
default we want to use the existing path to a program and not require
the Emacs user to set that path via customize functions.

This patch changes the default value of `love-exe` so that it will
first attempt to use system path to `love`.  Failing that we fall back
on the hard-coded path which seems to be to the benefit of Windows
developers, but since I do not use Windows I cannot test that idea.