[Style] Do not use a parentheses alone on a line in `love/play`
This patch does not affect the behavior of anything.  It simply makes
the function consistent with all existing code, which does not use
trailing parentheses to close s-expressions.  Some people prefer to
have such parentheses on their own lines, and others don't.  The point
of this patch is *not* to claim one approach is better.  Instead it
exists for a far more important reason: the style of code in a project
should be as consistent as possible, for the sake of readability.